Templates out the BuildCanvasBatchSystem.
