+++ Core/Components/Collections/Runtime/Bin.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c632cde7b33a420d8481b73c732d7f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
