+++ Core/Systems/Conversions/MaterialBinConversionSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e7f76533537054b53b790a7c6e6ae2d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
